Wizard refactor
