+++ src/main/scala/com/fortysevendeg/translatebubble/ui/wizard/Layout.scala
@@ -17,14 +17,20 @@
 package com.fortysevendeg.translatebubble.ui.wizard
 
 import android.content.Intent
+import android.support.v4.app.{FragmentManager, Fragment}
 import android.support.v4.view.ViewPager
+import android.support.v4.view.ViewPager.OnPageChangeListener
 import android.widget._
 import com.fortysevendeg.macroid.extras.ViewTweaks._
+import com.fortysevendeg.macroid.extras.ViewPagerTweaks._
+import com.fortysevendeg.macroid.extras.ViewGroupTweaks._
 import com.fortysevendeg.translatebubble.modules.persistent.PersistentServicesComponent
 import com.fortysevendeg.translatebubble.ui.preferences.MainActivity
 import com.fortysevendeg.translatebubble.ui.wizard.Styles._
 import macroid.FullDsl._
-import macroid.{ActivityContextWrapper, IdGeneration, Ui}
+import macroid._
+
+import scala.concurrent.ExecutionContext.Implicits.global
 
 trait Layout
   extends IdGeneration {
@@ -60,6 +66,42 @@ trait Layout
   def pagination(position: Int)(implicit contextWrapper: ActivityContextWrapper) = getUi(
     w[ImageView] <~ paginationItemStyle <~ vTag("position_%d".format(position))
   )
+
+  def initializeUi(modeTutorial: Boolean)(implicit contextWrapper: ActivityContextWrapper, managerContext: FragmentManagerContext[Fragment, FragmentManager]) = {
+    val stepsCount = Steps.steps.length
+    (viewPager <~
+      vpAdapter(new StepsPagerAdapter(managerContext.manager)) <~
+      vpOnPageChangeListener(new OnPageChangeListener {
+        var isLastStep = false
+
+        override def onPageScrollStateChanged(i: Int): Unit = {}
+
+        override def onPageScrolled(i: Int, v: Float, i1: Int): Unit = {}
+
+        override def onPageSelected(i: Int): Unit = {
+          val ui = (modeTutorial, isLastStep, stepsCount) match {
+            case (tutorial, _, steps) if !tutorial && i >= steps - 1 =>
+              isLastStep = true
+              (paginationContent <~~ (vGone ++ fadeOut(300))) ~
+                (gotIt <~ vVisible <~~ fadeIn(300))
+            case (tutorial, lastStep, _) if !tutorial && lastStep =>
+              isLastStep = false
+              (paginationContent <~ vVisible <~~ fadeIn(300)) ~
+                (gotIt <~~ (vGone ++ fadeOut(300)))
+            case _ => Ui.nop
+          }
+          runUi((paginationContent <~ activateImages(i)) ~ ui)
+        }
+      })) ~
+      (paginationContent <~
+        vgAddViews(0 until stepsCount map pagination) <~
+        activateImages(0))
+  }
+
+  def activateImages(position: Int) = Transformer {
+    case i: ImageView if i.getTag.equals("position_%d".format(position)) => i <~ vActivated(true)
+    case i: ImageView => i <~ vActivated(false)
+  }
 
 }
