+++ src/main/scala/com/fortysevendeg/translatebubble/ui/wizard/WizardActivity.scala
@@ -19,17 +19,11 @@ package com.fortysevendeg.translatebubble.ui.wizard
 import android.content.Intent
 import android.os.Bundle
 import android.support.v4.app._
-import android.support.v4.view.ViewPager.OnPageChangeListener
-import android.widget.ImageView
-import com.fortysevendeg.macroid.extras.ViewTweaks._
-import com.fortysevendeg.translatebubble.R
 import com.fortysevendeg.translatebubble.modules.ComponentRegistryImpl
 import com.fortysevendeg.translatebubble.ui.commons.Strings._
 import com.fortysevendeg.translatebubble.ui.preferences.MainActivity
 import macroid.FullDsl._
-import macroid.{ContextWrapper, Contexts, Transformer}
-
-import scala.concurrent.ExecutionContext.Implicits.global
+import macroid.{ContextWrapper, Contexts}
 
 class WizardActivity
   extends FragmentActivity
@@ -47,73 +41,9 @@ class WizardActivity
       startActivity(new Intent(this, classOf[MainActivity]))
       finish()
     }
-
     analyticsServices.send(if (modeTutorial) analyticsTutorialScreen else analyticsWizardScreen)
-
     setContentView(layout)
-
-    val steps = Steps.steps.length
-
-    for {
-      page <- paginationContent
-      pager <- viewPager
-    } yield {
-      def activateImages(position: Int) = Transformer {
-        case i: ImageView if i.getTag.equals("position_%d".format(position)) => i <~ vActivated(true)
-        case i: ImageView => i <~ vActivated(false)
-      }
-
-      pager.setAdapter(new StepsPagerAdapter(getSupportFragmentManager))
-      pager.setOnPageChangeListener(new OnPageChangeListener {
-        var isLastStep = false
-
-        override def onPageScrollStateChanged(i: Int): Unit = {}
-
-        override def onPageScrolled(i: Int, v: Float, i1: Int): Unit = {}
-
-        override def onPageSelected(i: Int): Unit = {
-          runUi(paginationContent <~ activateImages(i))
-          if (!modeTutorial) {
-            if (i >= steps - 1) {
-              isLastStep = true
-              runUi(
-                (paginationContent <~~ (vGone ++ fadeOut(300))) ~
-                  (gotIt <~ vVisible <~~ fadeIn(300))
-              )
-            } else if (isLastStep) {
-              isLastStep = false
-              runUi(
-                (paginationContent <~ vVisible <~~ fadeIn(300)) ~
-                  (gotIt <~~ (vGone ++ fadeOut(300)))
-              )
-            }
-          }
-        }
-      })
-      for (p <- 0 to (steps - 1)) {
-        page.addView(pagination(p))
-      }
-      runUi(paginationContent <~ activateImages(0))
-    }
-
-  }
-
-  class StepsPagerAdapter(fragmentManager: FragmentManager)
-    extends FragmentPagerAdapter(fragmentManager) {
-
-    val steps = Steps.steps
-
-    override def getItem(position: Int): Fragment = {
-      val fragment = new StepFragment()
-      val bundle = new Bundle()
-      bundle.putInt(StepFragment.keyStepPosition, position)
-      fragment.setArguments(bundle)
-      fragment
-    }
-
-    override def getCount = {
-      steps.length
-    }
+    runUi(initializeUi(modeTutorial))
   }
 
 }
@@ -122,31 +52,4 @@ object WizardActivity {
   val keyModeTutorial = "mode_tutorial"
 }
 
-object Steps {
 
-  def steps(implicit contextWrapper: ContextWrapper) = List(
-    Step(
-      R.drawable.wizard_icon,
-      contextWrapper.application.getString(R.string.wizardTitle1),
-      contextWrapper.application.getString(R.string.wizardDescription1)),
-    Step(
-      R.drawable.wizard_step_01,
-      contextWrapper.application.getString(R.string.wizardTitle2),
-      contextWrapper.application.getString(R.string.wizardDescription2)),
-    Step(
-      R.drawable.wizard_step_02,
-      contextWrapper.application.getString(R.string.wizardTitle3),
-      contextWrapper.application.getString(R.string.wizardDescription3)),
-    Step(
-      R.drawable.wizard_step_03,
-      contextWrapper.application.getString(R.string.wizardTitle4),
-      contextWrapper.application.getString(R.string.wizardDescription4)),
-    Step(
-      R.drawable.wizard_step_04,
-      contextWrapper.application.getString(R.string.wizardTitle5),
-      contextWrapper.application.getString(R.string.wizardDescription5)))
-
-
-}
-
-case class Step(image: Int, title: String, description: String)
